requirements.txt
appdirs==1.4.3
beautifulsoup4==4.9.3
bs4==0.0.1
CacheControl==0.12.6
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.1
contextlib2==0.6.0
crayons==0.4.0
dateparser==1.0.0
distlib==0.3.0
distro==1.4.0
fake-useragent==0.1.11
html5lib==1.0.1
htmldate==0.8.1
idna==2.10
ipaddr==2.2.0
lockfile==0.12.2
lxml==4.6.3
msgpack==0.6.2
numpy==1.19.4
packaging==20.3
pandas==1.1.5
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2020.4
regex==2021.3.17
requests==2.25.1
retrying==1.3.3
selenium==3.141.0
six==1.15.0
soupsieve==2.1
tzlocal==2.1
urllib3==1.26.2
webdriver-manager==3.2.2
webencodings==0.5.1
